Set a MainActivity to control the display of a user_home and advisor_home screen
